Fix: creation of placeholder files
Previously, as we utilised git trees to move files, the creation of the placeholder file was being overridden due to the git tree used being the original at the time of the request. The creation of the placeholder file has been shifted to after the git tree operation to resolve this.