fix(site launch): preventing race condition when 2 sites launched...
at the same time. Originally, LaunchesService had instance variables to
cache results from the db, rather than keep re-pining the db for the
siteId + appId. But, this might lead to a race condition if two forms
were submitted back-to-back. As such, data is passed back to the parent
to be reused, rather than storing it within the singular object
 being created during runtime.