feat(axios): implement etag functionality (#765)
* chore(axios): upgrade axios to 0.24 and add axios cache interceptor

pin axios to only allow minor patch upgrades, and pin the version of interceptor exactly. this is
because there is a support table
[here](https://axios-cache-interceptor.js.org/guide/getting-started) that lists the compat and we
require at least 0.6.x for ETag support. in order to prevent versioning errors (0.8.4 of interceptor
works wtih 0.25 of axios!!), we will pin the exact version of interceptor + patch version of axios

* feat(axiosinstance): add initialisation code

* chore(package): lock versions

* fix(axiosinstance): update config for cache interceptor