feat(quickie): only gitfile should have quickie (#1042)
## Problem

Moving forward, since we have migrated most sites out of gh, we will only offer quickie for sites that are in ggs. 
As such, this pr removes the dependency of a no longer supported flow so as to allow for a simpler mental model. 
Additionally, there were weird abstractions bet GithubService + GithubCommitService -> removed the responsibilities to be isolated.

Test cases will be added in the leaf node of these prs, cicd will also be fixed upstream.