fix(linkChecker): bug fixes (#1187)
## Problem

Users have given feedback regarding link checker. 

1. Decoding issues 
The repo had the file 
`/images/Life @ JPJC/2023 Events/LTC2023/ltc photo 1.JPG`

User had sourced this in one of their pages using the path `/images/Life%20%40%20JPJC/2023%20Events/LTC2023/ltc%20photo%201.JPG`

In the live site, this is a vaild link.

When running 
```
const uri = '/images/Life%20%40%20JPJC/2023%20Events/LTC2023/ltc%20photo%201.JPG';
  console.log(decodeURIComponent(uri));
  console.log(decodeURI(uri));
```

we get the result 
```
> "/images/Life @ JPJC/2023 Events/LTC2023/ltc photo 1.JPG"
> "/images/Life %40 JPJC/2023 Events/LTC2023/ltc photo 1.JPG"
```

which is expected as the [documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/decodeURI#decodeuri_vs._decodeuricomponent) states 
> decodeURI() assumes the input is a full URI, so it does not decode characters that are part of the URI syntax.

since we are dealing with user supplied inputs, we have to make the least guarantees, and assume that the user supplied value might contain characters that are part of the URI syntax. 

2. Anchor links
We have an agency that used pattern of 
`<a id="blah"></a>` 
for their anchor links. To them, this is not a broken link and a false positive, which is understandable. Since this is something that the end citizen will not notice, I am opting to make a call to not report any anchor tags without any text. 

3. Image links
There exist dangling image tags in .md files, eg
`<img/>` 
and this has confused the user. Similar to the reasoning above, removing reporting of empty img srcs. 

4. Trivial permalink normalisation issues. eg leading quotation marks 
5. any resources _post page in cms was not rendered properly
6. Remove query params when checking for local references, amplify allows it and returns a 200

Other optimisations
Allow faster dev iteration by removing the push to the remote repository isomer-link-checker

this pr should make link checker stable for the upcoming 8 sites to be whitelisted after release

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [ ] Yes - this PR contains breaking changes
  - Details ...
- [X] No - this PR is backwards compatible with ALL of the following feature flags in this [doc](https://www.notion.so/opengov/Existing-feature-flags-518ad2cdc325420893a105e88c432be5)


## Tests

nil