chore(commitService): add logging (#1211)
## Problem

we lack observability here. we experienced a silent failure in terms of divergence, and not a clue on how to debug since we cannot replicate this + no unusual errors around that time. 
this is a last ditch effort to see if we can get more info on this issue, but personally am not too optimistic about this pr since by right the top level handlers should log errors