fix(dep): remove redundant deps (#1333)
## Problem

We are seeing weird commits like [this](https://github.com/isomerpages/isomercms-backend/pull/1314/files). I previously added them in by mistake, thought i removed it alr but turns out its not. this is a cleanup, doesnt make sense to mark it as both a dep and a dev dep at the same time.

## Solution

![Screenshot 2024-04-25 at 10.55.34 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/4JosFH65rhzwIvkZw2J6/ff53f9e5-0bac-4958-a327-65d61d86a168.png)

rm the dep, since its just a dep dep 

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [ ] Yes - this PR contains breaking changes
  - Details ...
- [X] No - this PR is backwards compatible with ALL of the following feature flags in this [doc](https://www.notion.so/opengov/Existing-feature-flags-518ad2cdc325420893a105e88c432be5)